Cambio de variable a float

+++ include/Calcular.h
@@ -5,11 +5,11 @@
 
 struct _variableValor{
     int b;
-    int valor;
+    float valor;
 };
 
 typedef struct _variableValor vValor;
 
 vValor pertenece (cola *c1,char op);
 void ordenar (cola *c1, cola *c);
-int calcular (cola *c);
+float calcular (cola *c);
